Return noop continuation when not async propagating
