+++ dd-java-agent/instrumentation/opentracing/api-0.31/src/test/groovy/OpenTracing31Test.groovy
@@ -28,6 +28,7 @@ import io.opentracing.util.GlobalTracer
 import spock.lang.Subject
 
 import static datadog.trace.agent.test.utils.TraceUtils.runUnderTrace
+import static datadog.trace.bootstrap.instrumentation.api.AgentTracer.noopContinuation
 import static datadog.trace.bootstrap.instrumentation.api.AgentTracer.setAsyncPropagationEnabled
 
 class OpenTracing31Test extends AgentTestRunner {
@@ -164,7 +165,7 @@ class OpenTracing31Test extends AgentTestRunner {
     span instanceof MutableSpan
     scope instanceof TraceScope
     !internalTracer.isAsyncPropagationEnabled()
-    (scope as TraceScope).capture() == null
+    (scope as TraceScope).capture() == noopContinuation()
     (tracer.scopeManager().active().span().delegate == span.delegate)
 
     when:
