+++ dd-java-agent/instrumentation/opentracing/api-0.32/src/test/groovy/OpenTracing32Test.groovy
@@ -33,6 +33,7 @@ import static datadog.trace.api.sampling.PrioritySampling.USER_KEEP
 import static datadog.trace.api.sampling.SamplingMechanism.AGENT_RATE
 import static datadog.trace.api.sampling.SamplingMechanism.DEFAULT
 import static datadog.trace.api.sampling.SamplingMechanism.MANUAL
+import static datadog.trace.bootstrap.instrumentation.api.AgentTracer.noopContinuation
 import static datadog.trace.bootstrap.instrumentation.api.AgentTracer.setAsyncPropagationEnabled
 
 class OpenTracing32Test extends AgentTestRunner {
@@ -174,7 +175,7 @@ class OpenTracing32Test extends AgentTestRunner {
     span instanceof MutableSpan
     scope instanceof TraceScope
     !internalTracer.isAsyncPropagationEnabled()
-    (scope as TraceScope).capture() == null
+    (scope as TraceScope).capture() == noopContinuation()
     (tracer.scopeManager().active().span().delegate == span.delegate)
 
     when:
