+++ internal-api/build.gradle
@@ -74,7 +74,6 @@ excludedClassesCoverage += [
   "datadog.trace.bootstrap.instrumentation.api.AgentTracer.NoopAgentHistogram",
   "datadog.trace.bootstrap.instrumentation.api.AgentTracer.NoopAgentPropagation",
   "datadog.trace.bootstrap.instrumentation.api.AgentTracer.NoopAgentTraceCollector",
-  "datadog.trace.bootstrap.instrumentation.api.AgentTracer.NoopContinuation",
   "datadog.trace.bootstrap.instrumentation.api.AgentTracer.NoopPathwayContext",
   "datadog.trace.bootstrap.instrumentation.api.AgentTracer.NoopTraceConfig",
   "datadog.trace.bootstrap.instrumentation.api.AgentTracer.NoopTracerAPI",
@@ -90,6 +89,7 @@ excludedClassesCoverage += [
   "datadog.trace.bootstrap.instrumentation.api.InternalSpanTypes",
   "datadog.trace.bootstrap.instrumentation.api.NoopAgentScope",
   "datadog.trace.bootstrap.instrumentation.api.NoopAgentSpan",
+  "datadog.trace.bootstrap.instrumentation.api.NoopContinuation",
   "datadog.trace.bootstrap.instrumentation.api.NoopScope",
   "datadog.trace.bootstrap.instrumentation.api.NoopSpan",
   "datadog.trace.bootstrap.instrumentation.api.NoopSpanContext",
