+++ internal-api/src/main/java/datadog/trace/bootstrap/instrumentation/api/NoopScope.java
@@ -17,7 +17,7 @@ final class NoopScope implements AgentScope {
 
   @Override
   public Continuation capture() {
-    return AgentTracer.NoopContinuation.INSTANCE;
+    return NoopContinuation.INSTANCE;
   }
 
   @Override
